api/boc/requirements.txt
Django==1.11.2
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
